+++ lua/neorg/modules/core/keybinds/module.lua
@@ -213,6 +213,14 @@ module.private = {
                         "<Plug>(neorg.dirman.new-note)",
                         opts = { desc = "[neorg] Create New Note" },
                     },
+
+                    -- Create a Table of Contents
+                    -- ^Table of Contents
+                    {
+                        "gO",
+                        "<cmd>Neorg toc<CR>",
+                        opts = { desc = "[neorg] Create Table of Contents" },
+                    },
                 },
             },
             norg = {
@@ -278,7 +286,7 @@ module.private = {
                     {
                         "<LocalLeader>ta",
                         "<Plug>(neorg.qol.todo-items.todo.task-ambiguous)",
-                        opts = { desc = "[neorg] Mark as Ambigous" },
+                        opts = { desc = "[neorg] Mark as Ambiguous" },
                     },
 
                     -- Switch the task under the cursor between a select few states